+++ test/runtests.jl
@@ -12,6 +12,7 @@ const TEST_MODE = get(ENV, "TEST_MODE", "default")
         include("unit/test_conservation.jl")
         include("unit/test_observables.jl")
         include("unit/test_thermostat.jl")
+        include("unit/test_checkpoint.jl")
     end
     if TEST_MODE in ("default", "all", "aqua")
         include("aqua.jl")